[#52][news] refactor: 서치 디티오 dip 위배된 것들 수정
