+++ news/src/main/java/com/newpick4u/news/news/presentation/NewsAPIController.java
@@ -6,7 +6,7 @@ import com.newpick4u.news.news.application.dto.response.NewsResponseDto;
 import com.newpick4u.news.news.application.dto.response.NewsSummaryDto;
 import com.newpick4u.news.news.application.dto.response.PageResponse;
 import com.newpick4u.news.news.application.usecase.NewsService;
-import com.newpick4u.news.news.domain.critria.NewsSearchCriteria;
+import com.newpick4u.news.news.application.dto.NewsSearchCriteria;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameters;
 import com.newpick4u.common.response.ApiResponse;
